Updated Dominator Snipes to have a To-Hit Buff slider (0% to 22%, only, can't do negative numbers on sliders). This did not include the Radioactive Assault as that set still needs its own sweep. Otherwise, all Dom Snipes are updated.
Also fixed a bug in Quick Form inherent (FastSnipe). It was not being created when taking Moonbeam, this is because the Power Name "Moonbeam" now has the Display Name "Engulfing Darkness". The Power Name "Death Shroud" is now the Display Name "Moon Beam". The bug had Quick Form generating when the user chose Engulfing Darkness, now I fixed it to generate when taking Moon Beam.