Peacebringer Update: Fixed Glinting Eye and Gleaming Bolt to use new conditionals and AttribModify to handle their redirects for being in human mode and shapeshift mode. Also fixed some number entries for these powers.
Fixed the Mutual Exclusivity between Incandescence, Bright Nova and White Dwarf. Set Incandescence to Skip Mutual Exclusivity (previously was set to Auto-Detoggle other powers), and set both Nova and Dwarf to Auto-Detoggle other powers (Kheldian_Group), where previously is was set to Skip Mutual Exclusivity.

Fixed the endurance cost of Quantum Acceleration. It was set to 1 end/sec, when it's supposed to be 0.65 end/sec (endurance cost of 0.325 every 0.5s).