Added all Regular Attuned Sentinel Ward power data to the Boost powersets. Gave them effects, etc.
Corrected Set_Bonus.Entomb_Proc and Set_Bonus.Gauntleted_Fist. Both had their absorb procs set for PvP only. This is an error in logic from the fact there is a requirement for the target to be a player...but that player is you. So it works in PvE and PvP.