Updated Corruptor Water Blast for the following 4 powers: Aqua Bolt, Hydro Blast, Water Burst, and Whirlpool.
Also updated the Pet.Whirlpool_Corruptor power.

Finally, I updated the AttribMod for Class Minion Pets to reflect Level 50 Melee_Debuff_Def should be -0.10, it was incorrectly set to -0.075. (Titled this Rev 3 and dated it 1/19/2021). HOWEVER, I can't get Whirlpool to show the debuff as -10%. It still shows -7.5%. So this will need to be looked at.