Fixed all Procs that also enhanced a power (Overwhelming, ATOs, Winters). A bug was causing the enhancement to also act as a player buff (so a 23% recharge enhancement also was giving a 23% recharge buff).
Fixed Preventive Medicine's and Entomb's proc to show +Absorb effects. Also, Entomb was set to only work in PvP, that's been fixed to Any.

Fixed Dominion's proc to show -Damage Debuff. Its magnitude was set to 0, instead of 1. *Correction* I see this was set to 0 mag because of a bug in Mids that thinks the debuff applies to self. I instead changed this to a -Damage(None) so it has no impact.

Fixed all Winter Set procs to show their PPM and proc info. All of these did not have the "Add Special Attribute" set up in the Enhancement Editor, so their proc effects wouldn't show. That's been updated. Some exceptions however, the -Recharge -Slow proc shows nothing because it would debuff the character if I give it non-zero values.